divide save_trans_from_jq_gm_skysoft into save_jq/gm|trans_gm/trans_ss
